services/parser/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "card-reader-parser"
version = "0.1.0"
description = "Card Reader parser service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "card-reader-core",
  "paddlepaddle>=3.1.0",
  "paddleocr>=3.1.0",
  "paddlex[ocr-core]>=3.1.0",
  "pillow>=10.4.0",
  "pydantic-settings>=2.5.2",
  "sqlmodel>=0.0.22",
  "watchfiles>=0.24.0"
]

[tool.uv.sources]
card-reader-core = { path = "../core", editable = true }

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["card_reader_parser*"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.6.9",
  "mypy>=1.11.2"
]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = ["src"]
